accessTokenの読み込み処理を共通化した
